+++ src/client/car.h
@@ -6,9 +6,9 @@
 class Car{
     private:
         uint8_t id;
-        uint32_t x_position;
-        uint32_t y_position;
-        uint32_t angle;
+        float x_position;
+        float y_position;
+        float angle;
         bool pointing_right;
         bool nitro_flag;
         uint8_t nitro_quantity;
